+++ app/Http/Controllers/firstController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\models\firstModel;
+use App\models\User;
+use Illuminate\Support\Facades\Auth;
 class firstController extends Controller
 {
     //
@@ -12,26 +14,43 @@ class firstController extends Controller
         return view("form");
     }
     public function successful(Request $myData){
+        $user = User::find(Auth::User()->id);
         $newData = new firstModel;
-        $newData->name = $myData->name;
+        $newData->username = $myData->name;
         $newData->content = $myData->content;
-        $newData->save();
+        $user = $user->todoItems()->save($newData);
         return redirect ("/home");
     }
-    public function delete($id){
-        \Log::info($id);
-        $tbd=firstModel::find($id);
-        $tbd->delete();
-        return redirect ("/home");
+    public function delete(Request $r){
+        $tbd=firstModel::find($r->item_id);
+        if($tbd->user_id == Auth::User()->id){
+            $tbd->delete();
+            return redirect ("/home");
+        }
+        else{
+            return redirect ("/home");
+        }
+        
     }
     public function edit($id){
-        return view("edit",["editData"=>firstModel::find($id)]);
+        $editdata = firstModel::find($id);
+        if($editdata->user_id == Auth::User()->id){
+            return view("edit",["editData"=>$editdata]);
+        }
+        else{
+            return redirect ("/home");
+        }
     }
     public function update(Request $myData){
-        $tbd=firstModel::find($myData->id);
-        $tbd->name = $myData->name;
-        $tbd->content = $myData->content;
-        $tbd->update();
-        return redirect ("/home");
+        $model = firstModel::find($myData->id);
+        if($model->user_id == Auth::User()->id){
+            $model->username = $myData->name;
+            $model->content = $myData->content;
+            $model->update();
+            return redirect ("/home");
+        }
+        else{
+            return redirect ("/home");
+        }     
     }
 }
